+++ pkg/server/app/errors.go
@@ -15,4 +15,8 @@ var (
 	ErrInvalidEnvVarName       = status.Errorf(codes.InvalidArgument, "Invalid Env Var Name")
 	ErrInvalidSecretName       = status.Errorf(codes.InvalidArgument, "Invalid Secret Name")
 	ErrInvalidActionForCronJob = status.Errorf(codes.InvalidArgument, "Invalid action for a cronjob app")
+	ErrMissingVirtualHost      = status.Errorf(
+		codes.InvalidArgument,
+		"Missing --vhost argument with the application domain",
+	)
 )